requirements.txt
Flask==2.3.3
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow-cpu==2.13.0
xgboost==1.7.6
imbalanced-learn==0.11.0
